warp nav in section to satisfy I.1
