+++ src/components/layout/Header.tsx
@@ -1,4 +1,4 @@
-import { Navigation } from "../Navigation";
+import { Navigation } from "../navigation/Navigation";
 
 export function Header() {
   return (
